Refactor: Move other 2 classes to file package
