+++ common/src/main/java/bisq/common/file/FileUtil.java
@@ -17,7 +17,6 @@
 
 package bisq.common.file;
 
-import bisq.common.storage.ResourceNotFoundException;
 import bisq.common.util.Utilities;
 
 import com.google.common.io.Files;
