+++ common/src/main/java/bisq/common/file/ResourceNotFoundException.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.common.storage;
+package bisq.common.file;
 
 public class ResourceNotFoundException extends Exception {
     public ResourceNotFoundException(String path) {
