+++ core/src/main/java/bisq/core/dao/node/explorer/ExportJsonFilesService.java
@@ -28,7 +28,7 @@ import bisq.core.dao.state.model.blockchain.TxType;
 
 import bisq.common.config.Config;
 import bisq.common.file.FileUtil;
-import bisq.common.storage.JsonFileManager;
+import bisq.common.file.JsonFileManager;
 import bisq.common.util.Utilities;
 
 import org.bitcoinj.core.Utils;
