+++ core/src/main/java/bisq/core/offer/OfferBookService.java
@@ -28,9 +28,9 @@ import bisq.network.p2p.storage.payload.ProtectedStorageEntry;
 
 import bisq.common.UserThread;
 import bisq.common.config.Config;
+import bisq.common.file.JsonFileManager;
 import bisq.common.handlers.ErrorMessageHandler;
 import bisq.common.handlers.ResultHandler;
-import bisq.common.storage.JsonFileManager;
 import bisq.common.util.Utilities;
 
 import javax.inject.Inject;
