+++ core/src/main/java/bisq/core/trade/DumpDelayedPayoutTx.java
@@ -18,7 +18,7 @@
 package bisq.core.trade;
 
 import bisq.common.config.Config;
-import bisq.common.storage.JsonFileManager;
+import bisq.common.file.JsonFileManager;
 import bisq.common.util.Utilities;
 
 import javax.inject.Inject;
